+++ module/Activity/src/Activity/Service/Signup.php
@@ -3,7 +3,9 @@
 namespace Activity\Service;
 
 use Activity\Model\Activity as ActivityModel;
+use Activity\Model\ActivitySignup;
 use Activity\Model\UserActivitySignup;
+use Activity\Model\ExternalActivitySignup;
 use Application\Service\AbstractAclService;
 use Decision\Model\Member;
 
@@ -36,6 +38,7 @@ class Signup extends AbstractAclService
      * Otherwise, it returns it in the available language.
      *
      * @param type $fields
+     * @param bool $external Whether the signup is external.
      * @return type
      * @throws \User\Permissions\NotAllowedException
      */
@@ -49,6 +52,13 @@ class Signup extends AbstractAclService
         }
         $form = new \Activity\Form\ActivitySignup();
         $form->initialiseForm($fields);
+        return $form;
+    }
+
+    public function getExternalForm($fields)
+    {
+        $form = new \Activity\Form\ActivitySignup();
+        $form->initialiseExternalForm($fields);
         return $form;
     }
 
@@ -150,7 +160,7 @@ class Signup extends AbstractAclService
     }
 
     /**
-     * Sign up  an activity with the specified field values.
+     * Sign a User up for an activity with the specified field values.
      *
      * @param ActivityModel $activity
      * @param array $fieldResults
@@ -163,16 +173,82 @@ class Signup extends AbstractAclService
                 $translator->translate('You need to be logged in to sign up for this activity')
             );
         }
-
-        $em = $this->getServiceManager()->get('Doctrine\ORM\EntityManager');
-
-        // Find the current user
         $user = $this->getServiceManager()->get('user_service_user')->getIdentity();
-
         $signup = new UserActivitySignup();
-        $signup->setActivity($activity);
         $signup->setUser($user);
+        $this->createSignup($signup, $activity, $fieldResults);
+    }
+
+    /**
+     * Sign an external user up for an activity, which the current user may admin.
+     *
+     * @param \Activity\Service\AcitivityModel $activity
+     * @param type $fullName
+     * @param type $email
+     * @param array $fieldResults
+     * @throws \User\Permissions\NotAllowedException
+     */
+    public function adminSignUp(ActivityModel $activity, $fullName, $email, array $fieldResults)
+    {
+        if (!($this->isAllowed('adminSignup', 'activity') ||
+                $this->isAllowed('adminSignup', $activity))) {
+            $translator = $this->getTranslator();
+            throw new \User\Permissions\NotAllowedException(
+                $translator->translate('You are not allowed to create signups for this activity')
+            );
+        }
+        $this->manualSignUp($activity, $fullName, $email, $fieldResults);
+    }
+
+    /**
+     * Sign an external user up for an activity, allowed by a guest.
+     *
+     * @param \Activity\Service\AcitivityModel $activity
+     * @param type $fullName
+     * @param type $email
+     * @param array $fieldResults
+     * @throws \User\Permissions\NotAllowedException
+     */
+    public function externalSignUp(ActivityModel $activity, $fullName, $email, array $fieldResults)
+    {
+        if (!($this->isAllowed('externalSignup', 'activitySignup'))) {
+            $translator = $this->getTranslator();
+            throw new \User\Permissions\NotAllowedException(
+                $translator->translate('You are not allowed to create signups for this activity')
+            );
+        }
+        $this->manualSignUp($activity, $fullName, $email, $fieldResults);
+    }
+
+    /**
+     * Sign an external user up for an activity.
+     *
+     * @param \Activity\Service\AcitivityModel $activity
+     * @param type $user
+     * @param array $fieldResults
+     * @throws \User\Permissions\NotAllowedException
+     */
+    protected function manualSignUp(ActivityModel $activity, $fullName, $email, array $fieldResults)
+    {
+        $signup = new ExternalActivitySignup();
+        $signup->setEmail($email);
+        $signup->setFullName($fullName);
+        $this->createSignup($signup, $activity, $fieldResults);
+    }
+
+    /**
+     * Creates the generic parts of a signup.
+     * @param ActivitySignup $signup
+     * @param ActivityModel $activity
+     * @param type $user
+     * @param array $fieldResults
+     * @return ActivitySignup
+     */
+    protected function createSignup(ActivitySignup $signup, ActivityModel $activity, array $fieldResults)
+    {
+        $signup->setActivity($activity);
         $optionMapper = $this->getServiceManager()->get('activity_mapper_activity_option');
+        $em = $this->getServiceManager()->get('Doctrine\ORM\EntityManager');
         foreach ($activity->getFields() as $field){
             $fieldValue = new \Activity\Model\ActivityFieldValue();
             $fieldValue->setField($field);
@@ -196,6 +272,7 @@ class Signup extends AbstractAclService
         }
         $em->persist($signup);
         $em->flush();
+        return $signup;
     }
 
     /**
@@ -220,13 +297,25 @@ class Signup extends AbstractAclService
         if (is_null($signUp)) {
             return;
         }
+        $this->removeSignUp($signUp);
+    }
 
-        $em = $this->getServiceManager()->get('Doctrine\ORM\EntityManager');
-        $values = $this->getActivityFieldValueMapper()->getFieldValuesBySignup($signUp);
-        foreach($values as $value){
-            $em->remove($value);
+    public function externalSignOff(ExternalActivitySignup $signup)
+    {
+        if (!($this->isAllowed('adminSignup', 'activity') ||
+                $this->isAllowed('adminSignup', $signup->getActivity()))) {
+            $translator = $this->getTranslator();
+            throw new \User\Permissions\NotAllowedException(
+                $translator->translate('You are not allowed to remove external signups for this activity')
+            );
         }
-        $em->remove($signUp);
+        $this->removeSignUp($signup);
+    }
+
+    protected function removeSignUp(ActivitySignup $signup)
+    {
+        $em = $this->getServiceManager()->get('Doctrine\ORM\EntityManager');
+        $em->remove($signup);
         $em->flush();
     }